Add Fission auth
Add Fission auth alongisde the current auth system. New buttons have been added
for sign up and sign in.
Authentication leverages the current token system by storing a second token in
the webnative file system (WNFS) when a user signs up or signs in. This token
is stored encrypted at rest. When a user signs out, their active token is
invalidated, but they have a second token to sign them back in.
Webnative and webnative-filecoin have been added to a useFissionAuth custom
hook. Most of the interactions with webnative are handled here, with the
exception of the initial sign up.