Redirect the user to account setup when no token
If the user is authed with Fission, but they are missing a stored token,
redirect them to the account setup page. In theory, at this point they should
have a stored token, but if it is missing they will a new account or some
support. Note that they should not lose their wallet in this situation because
the client half of their cosigner key is stored separately.